requirements.txt
Flask>=3.0.3
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Werkzeug>=3.0.3
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
gunicorn>=22.0.0
